[FLINK-16190][e2e] Move FlinkDistribution into flink package
